move dismissal to parent
